Chatbot-BE/requirements.txt
pypdf==4.0.1
pathlib2==2.3.7
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
requests==2.31.0 